+++ tests/integration/test_cmd_script.py
@@ -111,11 +111,13 @@ class CmdSuite(unittest.TestCase):
                 stdout=subprocess.DEVNULL,
             )
 
-        _ = subprocess.run(
+        process = subprocess.run(
             ["gamspy", "install", "solver", "minos", "mosek"],
-            stderr=subprocess.DEVNULL,
-            stdout=subprocess.DEVNULL,
+            capture_output=True,
+            text=True,
         )
+        print(process.stdout, process.stderr)
+        self.assertTrue(process.returncode == 0)
 
         with self.assertRaises(subprocess.CalledProcessError):
             _ = subprocess.run(
@@ -125,11 +127,36 @@ class CmdSuite(unittest.TestCase):
                 stdout=subprocess.DEVNULL,
             )
 
-        _ = subprocess.run(
+        process = subprocess.run(
+            ["gamspy", "install", "solver", "--install-all-solvers"],
+            capture_output=True,
+            text=True,
+        )
+        print(process.stdout, process.stderr)
+        self.assertTrue(process.returncode == 0)
+
+        process = subprocess.run(
             ["gamspy", "uninstall", "solver", "minos", "mosek"],
+            capture_output=True,
+            text=True,
+        )
+        print(process.stdout, process.stderr)
+        self.assertTrue(process.returncode == 0)
+
+        process = subprocess.run(
+            ["gamspy", "uninstall", "solver", "--uninstall-all-solvers"],
+            capture_output=True,
+            text=True,
+        )
+        print(process.stdout, process.stderr)
+        self.assertTrue(process.returncode == 0)
+
+        process = subprocess.run(
+            ["gamspy", "install", "solver", "mpsge", "scip"],
             stderr=subprocess.DEVNULL,
             stdout=subprocess.DEVNULL,
         )
+        self.assertTrue(process.returncode == 0)
 
     def test_list_solvers(self):
         process = subprocess.run(
@@ -212,7 +239,7 @@ class CmdSuite(unittest.TestCase):
         )
 
     @classmethod
-    def tearDownClass(cls):
+    def tearDown(cls):
         _ = subprocess.run(
             ["gamspy", "install", "license", os.environ["LOCAL_LICENSE"]],
             check=True,