+++ CHANGELOG.md
@@ -6,6 +6,8 @@ GAMSPy 1.0.1
   - Fix frozen solve with non-scalar symbols.
   - Fix the definition update problem while redefining an equation with definition argument.
   - Introduce default directories to keep license information on upgrade.
+  - Add --existing-solvers and --install-all-solvers options for gamspy install solver.
+  - Add --uninstall-all-solvers option for gamspy uninstall solver.
   - Show license path on gamspy show license command.
   - Simplify the implementation of the copy container operation.
 - Testing
@@ -14,6 +16,9 @@ GAMSPy 1.0.1
   - Test the usage of a license that is in one of the default paths.
   - Fix the issue related to reading equation records from a gdx file.
   - Add tests to verify the records after reading them from a gdx file.
+  - Add tests for installing/uninstalling solvers.
+- Documentation
+  - Update the documentation of install/uninstall command line arguments.
 
 GAMSPy 1.0.0
 ------------